fix(LimitQueryHandler.java): Support receiving Number as limit argument
Previously only String was supported.
Now you can pass a String or any type of Number, it will be converted
to Long.
If an invalid value is passed IllegalArgumentException will be thrown

Also added a first Java test class. The test setup is not completed in this repository
but the test will run when added to platforms/android/app module of a cordova project
with this plugin installed with JUnit4 and mockito test dependencies